raytracer v0.2

+++ src/tools/Intersectable.rs
@@ -3,12 +3,12 @@ use image::Rgb;
 use super::Vector3::Vector3;
 use crate::objects::Ray::Ray;
 
-#[derive(Clone)]
+#[derive(Clone, Debug)]
 pub struct Intersection {
     pub distance: f64,
     pub position: Vector3,
     pub normal: Vector3,
-    pub color: Rgb<f32>,
+    pub color: Rgb<u8>,
 }
 
 pub trait Intersectable {
